(docs) Add not-working WebComponents exploration - custom elements *require* ES6
This is really annoying but it seems that you can not define custom elements
with ES5 syntax without a lot of work in setting up the build chain with
transpiling, polyfilling and such. This also explains why `x-tag` now uses
ES6 classes - they simply have no choice. Postponing WC until we find a
reasonable solution that works in IE11 or until our baseline is not IE11 anymore.